move inital tests to appropriate package
