+++ src/hakunapi-jsonfg/src/test/java/fi/nls/hakunapi/jsonfg/JSONFGSingleFeatureWriterTest.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.geojson.hakuna;
+package fi.nls.hakunapi.jsonfg;
 
 import java.io.ByteArrayOutputStream;
 import java.util.Collections;
@@ -9,20 +9,19 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
-import fi.nls.hakunapi.core.FeatureCollectionWriter;
 import fi.nls.hakunapi.core.FeatureStream;
 import fi.nls.hakunapi.core.FloatingPointFormatter;
+import fi.nls.hakunapi.core.ValueProvider;
 import fi.nls.hakunapi.core.geom.HakunaGeometryDimension;
+import fi.nls.hakunapi.core.property.HakunaProperty;
 import fi.nls.hakunapi.core.request.GetFeatureCollection;
 import fi.nls.hakunapi.core.request.GetFeatureRequest;
 import fi.nls.hakunapi.core.util.DefaultFloatingPointFormatter;
-import fi.nls.hakunapi.jsonfg.JSONFGFeatureCollectionWriter;
-import fi.nls.hakunapi.jsonfg.JSONFGTestUtils;
 
-public class JSONFGFeatureCollectionWriterTest extends JSONFGTestUtils {
+public class JSONFGSingleFeatureWriterTest extends JSONFGTestUtils {
 
 	@Test
-	public void testWriteJSONFGFeatureCollection() throws Exception {
+	public void testWriteJSONFGFeature() throws Exception {
 		FloatingPointFormatter f = new DefaultFloatingPointFormatter(0, 5, 0, 5, 0, 13);
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 
@@ -31,18 +30,25 @@ public class JSONFGFeatureCollectionWriterTest extends JSONFGTestUtils {
 		request.addCollection(coll);
 
 		try (FeatureStream fs = new TestFeaturesStream();
-				FeatureCollectionWriter fw = new JSONFGFeatureCollectionWriter()) {
+				JSONFGSingleFeatureWriter fw = new JSONFGSingleFeatureWriter()) {
+
+			ValueProvider feat = fs.next();
 
 			fw.init(baos, f, PLACE_SRID);
 			fw.initGeometryWriter(HakunaGeometryDimension.XY);
 
-			fw.startFeatureCollection(ft, null);
+			// Note: contains implicit call to startFeature !
+			int i = 0;
+			for (HakunaProperty prop : coll.getProperties()) {
+				prop.write(feat, i++, fw);
+			}
 
-			int numberReturned = writeFeatureCollection(fw, ft, coll.getProperties(), fs, 0, -1).numberReturned;
-
-			fw.endFeatureCollection();
-			fw.end(false, Collections.emptyList(), numberReturned);
+			fw.endFeature();
+			fw.end(true, Collections.emptyList(), 1);
 		}
+
+		System.out.println(baos.toString("UTF-8"));
+		System.out.flush();
 
 		ObjectMapper mapper = new ObjectMapper().enable(SerializationFeature.INDENT_OUTPUT)
 				.disable(SerializationFeature.CLOSE_CLOSEABLE);
